requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
pyyaml==6.0.1